feat: tooltips components
